Don't add properties for consul folders
The Consul KeyValue client returns folder names as KeyValue pairs
with null values. For example, if there is a config/application/ldap/
folder, then the KeyValue client returns this as a KeyValue and then
spring-cloud-consul adds this as a property, eg
config.application.ldap.=null. This is not a real property and
shouldn't be added to the environment. Specifically, this causes the
bean binding phase to fail since it can think it has a property with
a zero-length name if it finds a @ConfigurationProperties bean with
the prefix ldap (for example).