adds additional test for consul-discovery with ACL
this commit also performs little changes on the ACL setup. Giving "anonymous" write rights on all services is higher rated then the limitation of the discovery client acl. With the current setup now the additional test would fail, if the discovery client would not be aware of configured consul tokens.