Support different format for defaultContext
Allow the format for the defaultContext to be configured via
ConsulConfigProperties#defaultContextFormat.

Since this can create a mixture of formats among the contexts
loaded from ConsulPropertySourceLocator we need to keep track
of the format for each context when adding suffixes, and resolving
profile separator, and when loading the actual data from consul.
Because of this the ConsulPropertyContext class has been added.

Fixes gh-637